Break apart the manage tokens styles
