+++ src/components/common/ManageTokenRow.js
@@ -2,13 +2,14 @@
 
 import type { EdgeMetaToken } from 'edge-core-js'
 import React, { Component } from 'react'
-import { TouchableHighlight, TouchableWithoutFeedback, View } from 'react-native'
+import { StyleSheet, TouchableHighlight, TouchableWithoutFeedback, View } from 'react-native'
 import Icon from 'react-native-vector-icons/SimpleLineIcons'
 
 import { SYNCED_ACCOUNT_DEFAULTS } from '../../modules/Core/Account/settings.js'
 import CheckBox from '../../modules/UI/components/CheckBox/CheckBox.ui.js'
 import Text from '../../modules/UI/components/FormattedText/FormattedText.ui.js'
-import { styles, underlayColor } from '../../styles/scenes/ManageTokensStyle.js'
+import { THEME } from '../../theme/variables/airbitz.js'
+import { scale } from '../../util/scaling.js'
 import * as UTILS from '../../util/utils.js'
 
 export type State = {
@@ -46,7 +47,7 @@ class ManageTokenRow extends Component<Props, State> {
     const onPress = isEditable ? goToEditTokenScene : UTILS.noOp
 
     return (
-      <TouchableHighlight onPress={() => onPress(item.currencyCode)} underlayColor={underlayColor} style={styles.manageTokenRow}>
+      <TouchableHighlight onPress={() => onPress(item.currencyCode)} underlayColor={THEME.COLORS.PRIMARY_BUTTON_TOUCHED} style={styles.manageTokenRow}>
         <View style={styles.manageTokenRowInterior}>
           <View style={styles.rowLeftArea}>
             <TouchableWithoutFeedback onPress={() => toggleToken(item.currencyCode)} isVisible={item.isVisible} enabled={enabled}>
@@ -66,5 +67,45 @@ class ManageTokenRow extends Component<Props, State> {
     )
   }
 }
+
+const rawStyles = {
+  manageTokenRow: {
+    height: scale(44),
+    flexDirection: 'row',
+    borderBottomWidth: 1,
+    borderBottomColor: THEME.COLORS.GRAY_3,
+    paddingLeft: scale(20),
+    paddingRight: scale(20)
+  },
+  rowLeftArea: {
+    flexDirection: 'row'
+  },
+  manageTokenRowInterior: {
+    flex: 1,
+    flexDirection: 'row',
+    justifyContent: 'space-between',
+    alignItems: 'center'
+  },
+  touchableCheckboxInterior: {
+    paddingHorizontal: scale(8),
+    justifyContent: 'center',
+    alignItems: 'center'
+  },
+  checkBox: {
+    alignSelf: 'center'
+  },
+  tokenNameArea: {
+    alignSelf: 'center'
+  },
+  tokenNameText: {
+    color: THEME.COLORS.GRAY_1,
+    fontSize: scale(16)
+  },
+  rowRightArrow: {
+    fontSize: scale(18),
+    color: THEME.COLORS.GRAY_1
+  }
+}
+const styles: typeof rawStyles = StyleSheet.create(rawStyles)
 
 export default ManageTokenRow
